feat: Pass region in signup request body
Previously we used the region passed by the user on the command line to determine which url to use for the signup request. However, the identity stack is now global and takes the region as part of the body for the create token request. This updates to pass the region in the body rather than using it to determine the url (which is just hardcoded inline for now).