fix: Pass signup url as env var rather than flag
We don't want the signup url override to be an obvious thing for users - its mostly for dev and testing purposes. There doesn't seem to be an easy way to not have help text for a flag displayed, so instead make it an env var so it doesn't show up in the help output.