fix: Update help text for --region
Update help text for --region flag for account signup action to list out the currently available regions. This list will need to be updated as we add regions, but since it's in the help text it won't block users from using new regions before this list gets updated or without updating the cli.