feat: Update signup flow for global identity stack (#55)
Previously for the `account signup` command, the region passed by the `--region` flag determined which url to use for the signup request. However, the signup stack is now global and takes the region as part of the body for the create token request. This updates the command to pass the region in the body rather than using it to determine the url. This also talks to a new endpoint for signup.

Note: Old versions of the CLI will stop working for the `account signup` command.