feat: add `login` command (#152)
This is an active construction zone as we build out the capability.
Looking forward to when we have `release` and `nightly` distribution
channels.

Adds `momento login` command for doing SSO. The integration has
a live connection back up to the Momento endpoint that it running
the login workflow.

A standard oauth dance happens, with a browser opening up to an
SSO page. Some infos are printed from the server while you're
logging in, then the login credentials are printed to stdout.

We will be moving this session token to a momento dotfile, but
this change gets the SSO integration started. Not many endpoints
are supported, but over time we'll add more.