fix: broken brew update command in README
The command `brew update momento-cli` fails with an error: 
```
Error: This command updates brew itself, and does not take formula names.
Use `brew upgrade momento-cli` instead.
```

I think we need just `brew update`.  I don't like this because it potentially touches a bunch of stuff that a user isn't trying to upgrade, but it won't work if you pass it a formula name, and I don't know how to update a single formula, and I don't think the `brew upgrade momento-cli` command actually works without the `update` before it?