chore: only use logging when verbose
This turns off logging when not verbose. Verbose logging is set to
`debug` level. As discussed in #172 we find the log data _too_ verbose
compared to the other status messages or normal cache response data.

We replace other print statements with one of two new macros.

We funnel all status or error print statements to `stderr` with
`console_info!`. Similarly we define a macro `console_data!` for cache
response data, which we print to `stdout`.

Closes #172