chore: refactor to remove FileTypes enum
Prior to this commit, a great many functions were implemented
with a match statement that branched on Config vs. Credential
files.  This repetition of the match statements meant that
it was hard to reason about the logic for managing config
vs. credentials, and caused separation of concerns issues
for all of these functions.

We can follow up with some additional refactors to DRY up
the handful of places where this leaves some repetitive logic
behind, but this is a step forward in that now all of the
functions have a clear focus on either config or credentials.