chore: remove unnecessary conditional logic from update_*_profile fns
Prior to this commit there was a conditional in each of the
update_*_profile fns that branched on whether the current
iteration of the loop was the first one or not.  This conditional
is not necessary if we just clone the file content vector up
front; this commit does that.  Still not efficient yet but the
logic is simplified.