chore: simplify replace_*_value fns
The replace_*_value fns are responsible for replacing a config
or credentials value for a single line in the config files.  However,
prior to this commit, they were making a full copy of the file
contents in order to achieve it, as well as re-compiling the regexes
on every call.

This commit moves the regexes to static refs that are compiled
once on the initial load, and changes the signatures of the
replace_*_value fns so that they just accept a single line as
input and operate on that line rather than making a copy of the
entire file content.