fix: add protoc for aarch64 cross build
this docker image uses x86_64 to cross compile to aarch64, even on
aarch64 hosts, so x64_64 is the right protoc version.

Scripts aren't conveniently on the build path or I would have invoked
the script. However, we need to remove this protoc dependency from our
client_protos package anyway so I'm not going to over-engineer this