chore: introduce separate lib and app for binary vs. cli options
We have had a customer request for tab completion in the CLI.  In
order to support that we need to move the CLI options into their
own library that can be referenced via build.rs from the main app.

This commit introduces the new directory structure.  No files are
actually moved yet.