feat: hide signing key command from help, improve preview text
This commit moves the signing key command to "preview" status,
so that it doesn't show up in the default help.  It will still
show up in tab completion at the root of the command, but
is now prefixed with our "**PREVIEW**" label, and includes some
information in the subcommand help text about its preview status.