chore: improve progress bars in cloud-linter
Separate the metrics calls from the describe tables calls for
elasticache and serverless elasticache. This lets us use a spinner for
the describe calls, and then a standard progress bar for the metrics
calls.

Remove an extra call to increment the s3 progress bar.

Add a progress bar to display that we are writing out the results file.

Misc linter fixes.