feat: add cloud linter options to filter by elasticache engine
Add new cloud linter flags: elasti-cache-redis, elasti-cache-memcached,
and elasti-cache-serverless, in addition to the existing elasti-cache
flag. These limit the cloud linter to only process elasticache clusters
with the redis or memcached engines, or only serverless clusters.

Add a stub section for valkey elasticache clusters, to be filled out in
the future.

Make the unknown engine option return an empty vec instead of an error,
because the previously unknown valkey engine was breaking processing
when encountered.